+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - **SwiftLint** Delete unchanged files [#1069](https://github.com/sider/runners/pull/1069)
 - Log issues count [#1064](https://github.com/sider/runners/pull/1064)
 - Ensure `Processor#analyzer` initialization [#1067](https://github.com/sider/runners/pull/1067)
+- Add `analyzer` field to `Results::Error` [#1068](https://github.com/sider/runners/pull/1068)
 
 ## 0.23.0